Feat/multi client

+++ scripts/backup_database.py
@@ -4,7 +4,9 @@ import tarfile
 import json
 import boto3
 import shutil
+import argparse
 from logger import log_message
+from datetime import datetime
 
 def create_cnf_file(db_config):
     cnf_content = f"[client]\nuser = {db_config['user']}\npassword = {db_config['password']}\nhost = {db_config['host']}\nport = {db_config['port']}"
@@ -24,12 +26,15 @@ def backup_database(db_config, storage_config, temp_dir, timestamp):
     db_port = str(db_config["port"])
     db_type = db_config.get("type", "postgres")
 
-    s3_client = boto3.client(
-        "s3",
-        endpoint_url=storage_config["endpoint_url"],
-        aws_access_key_id=storage_config["access_key"],
-        aws_secret_access_key=storage_config["secret_key"],
-    )
+    # Conditionally create s3_client only if storage type is not "local"
+    s3_client = None
+    if storage_config.get("type") != "local":
+        s3_client = boto3.client(
+            "s3",
+            endpoint_url=storage_config["endpoint_url"],
+            aws_access_key_id=storage_config["access_key"],
+            aws_secret_access_key=storage_config["secret_key"],
+        )
 
     log_message(f"Initiating backup of {db_type} database {db_name}...")
 
@@ -72,18 +77,21 @@ def backup_database(db_config, storage_config, temp_dir, timestamp):
         with tarfile.open(tar_path, "w:gz") as tar:
             tar.add(dump_path, arcname=os.path.basename(dump_path))
 
-        storage_path = f"{timestamp}/database/{db_name}_backup.tar.gz"
-        log_message(f"Uploading to path: {storage_path}")
-
-        s3_client.upload_file(Bucket=storage_config["bucket_name"], Key=storage_path, Filename=tar_path)
-        log_message(f"Backup of database {db_name} completed.")
+        if storage_config.get("type") == "local":
+            log_message(f"Backup of database {db_name} saved locally at {temp_dir}.")
+        else:
+            storage_path = f"{timestamp}/database/{db_name}_backup.tar.gz"
+            log_message(f"Uploading to path: {storage_path}")
+            s3_client.upload_file(Bucket=storage_config["bucket_name"], Key=storage_path, Filename=tar_path)
+            log_message(f"Backup of database {db_name} completed.")
 
     except subprocess.CalledProcessError as e:
         log_message(f"Error executing database dump of {db_name}: {e}")
     except Exception as e:
         log_message(f"General error in database backup of {db_name}: {e}")
     finally:
-        shutil.rmtree(temp_dir)
+        if storage_config.get("type") != "local":
+            shutil.rmtree(temp_dir)
         if cnf_path:
             os.remove(cnf_path)
 
@@ -101,5 +109,9 @@ def backup_databases(config_path, timestamp):
 
 
 if __name__ == "__main__":
-    print("You shouldn't run this file directly.")
-    pass
+    parser = argparse.ArgumentParser()
+    parser.add_argument("--config", default="config/settings.json", help="Path to the configuration file")
+    args = parser.parse_args()
+
+    timestamp = datetime.now().strftime("%Y-%m-%dT%H:%M:%S")
+    backup_databases(args.config, timestamp)
